Add files via upload
This is a basic todo list web app done by using HTML, CSS and JavaScript. In this we can add our work, remove, edit it, while adding the task's the value of the tasks get increase by one but initially it is zero. Click on your task if it is completed it would strike off the task and decrease the value in the pending task this is how it works.